# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Account:
    properties:
      account_number:
        type: string
      account_number_aliases:
        type: string
      link_for_domain_admin:
        type: string
      name:
        type: string
      user_name:
        type: string
    type: object
...
  models.CreateAccount:
    properties:
      account_number:
        type: string
      account_number_aliases:
        type: string
      link_for_domain_admin:
        type: string
      name:
        type: string
      user_name:
        type: string
    type: object
...
  models.CreateCustomer:
    properties:
      backup_software:
        type: string
      cloud_or_onsite_id:
        type: integer
      created_by_user_id:
        type: integer
      description:
        type: string
      firewall:
        type: boolean
      hardware_as_a_service:
        type: boolean
      internet:
        type: boolean
      is_active:
        type: boolean
      name:
        type: string
      voip:
        type: boolean
    type: object
...
  models.CreateCustomerLocation:
    properties:
      created_by_user_id:
        type: integer
      customer_id:
        type: integer
      description:
        type: string
      is_primary:
        type: boolean
      name:
        type: string
    type: object
  models.CreateEmailAccount:
    properties:
      account_number:
        type: string
      email_account_type_id:
        type: integer
      email_hosting:
        type: string
      link_for_email_admin:
        type: string
      password:
        type: string
      pin:
        type: string
      user_name:
        type: string
    type: object
  models.CreateEmailDomain:
    properties:
      a_record_1:
        type: string
      a_record_2:
        type: string
      a_record_3:
        type: string
      a_record_4:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      domain:
        type: string
      domain_registrar:
        type: string
      mx_record_1:
        type: string
      mx_record_2:
        type: string
      notes:
        type: string
      password:
        type: string
      phone_setting_note:
        type: string
      pin:
        type: string
      web_mail_exchange:
        type: string
      web_mail_pop_imap:
        type: string
      website_ip_or_alias:
        type: string
    type: object
...
  models.Customer:
    properties:
      backup_software:
        type: string
      cloud_or_onsite_id:
        type: integer
      created_at:
        type: string
      created_by_user_id:
        type: integer
      customer_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      description:
        type: string
      firewall:
        type: boolean
      hardware_as_a_service:
        type: boolean
      internet:
        type: boolean
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
      voip:
        type: boolean
    type: object
...
  models.CustomerLocation:
    properties:
      created_at:
        type: string
      created_by_user_id:
        type: integer
      customer_id:
        type: integer
      customer_location_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      description:
        type: string
      is_primary:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
    type: object
  models.EmailAccount:
    properties:
      account_number:
        type: string
      email_account_type:
        $ref: '#/definitions/models.EmailAccountType'
      email_account_type_id:
        type: integer
      email_hosting:
        type: string
      link_for_email_admin:
        type: string
      password:
        type: string
      pin:
        type: string
      user_name:
        type: string
    type: object
  models.EmailAccountType:
    properties:
      email_account_type:
        type: string
      email_account_type_id:
        type: integer
    type: object
  models.EmailDomain:
    properties:
      a_record_1:
        type: string
      a_record_2:
        type: string
      a_record_3:
        type: string
      a_record_4:
        type: string
      account:
        $ref: '#/definitions/models.Account'
      account_id:
        type: integer
      created_at:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      domain:
        type: string
      domain_registrar:
        type: string
      email_account:
        $ref: '#/definitions/models.EmailAccount'
      email_account_id:
        type: integer
      email_domain_id:
        type: integer
      mx_record_1:
        type: string
      mx_record_2:
        type: string
      notes:
        type: string
      password:
        type: string
      phone_setting_note:
        type: string
      pin:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
      web_mail_exchange:
        type: string
      web_mail_pop_imap:
        type: string
      website_ip_or_alias:
        type: string
    type: object
...
  models.UpdateAccount:
    properties:
      account_number:
        type: string
      account_number_aliases:
        type: string
      link_for_domain_admin:
        type: string
      name:
        type: string
      user_name:
        type: string
    type: object
...
  models.UpdateCustomer:
    properties:
      backup_software:
        type: string
      cloud_or_onsite_id:
        type: integer
      description:
        type: string
      firewall:
        type: boolean
      hardware_as_a_service:
        type: boolean
      internet:
        type: boolean
      is_active:
        type: boolean
      name:
        type: string
      updated_by_user_id:
        type: integer
      voip:
        type: boolean
    type: object
...
  models.UpdateCustomerLocation:
    properties:
      customer_id:
        type: integer
      description:
        type: string
      is_primary:
        type: boolean
      name:
        type: string
      updated_by_user_id:
        type: integer
    type: object
  models.UpdateEmailAccount:
    properties:
      account_number:
        type: string
      email_account_type_id:
        type: integer
      email_hosting:
        type: string
      link_for_email_admin:
        type: string
      password:
        type: string
      pin:
        type: string
      user_name:
        type: string
    type: object
  models.UpdateEmailDomain:
    properties:
      a_record_1:
        type: string
      a_record_2:
        type: string
      a_record_3:
        type: string
      a_record_4:
        type: string
      customer_location_id:
        type: integer
      domain:
        type: string
      domain_registrar:
        type: string
      mx_record_1:
        type: string
      mx_record_2:
        type: string
      notes:
        type: string
      password:
        type: string
      phone_setting_note:
        type: string
      pin:
        type: string
      updated_by_user_id:
        type: integer
      web_mail_exchange:
        type: string
      web_mail_pop_imap:
        type: string
      website_ip_or_alias:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server .
  termsOfService: http://swagger.io/terms/
  title: Network Management Solutions API
  version: "1.0"
paths:
...
  /customer-locations:
    delete:
      description: Delete Customer Location
      parameters:
      - description: Customer Location ID
        in: query
        name: customer_location_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerLocation'
      summary: Delete Customer Location
      tags:
      - Customer Location
    get:
      description: Retrieve Customer Location
      parameters:
      - description: Customer Location ID
        in: query
        name: customer_location_id
        type: string
      - description: Customer Location Name
        in: query
        name: customer_location_name
        type: string
      - description: Customer Name
        in: query
        name: customer_name
        type: string
      - description: Assign To
        in: query
        name: assigned_to
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order_by
        type: string
      - description: Select Column
        in: query
        name: select_column
        type: string
      - description: Set Limit
        in: query
        name: set_limit
        type: string
      - description: Append Select
        in: query
        name: append_select
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerLocation'
      summary: Retrieve Customer Location
      tags:
      - Customer Location
    patch:
      consumes:
      - application/json
      description: Update Customer Location
      parameters:
      - description: CustomerLan Data
        in: body
        name: CustomerLocation
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCustomerLocation'
      - description: Customer Location ID
        in: query
        name: customer_location_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerLocation'
      summary: Update Customer Location
      tags:
      - Customer Location
    post:
      consumes:
      - application/json
      description: Add    Customer Location
      parameters:
      - description: CustomerLocation Data
        in: body
        name: CustomerLocation
        required: true
        schema:
          $ref: '#/definitions/models.CreateCustomerLocation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CustomerLocation'
      summary: Create Customer Location
      tags:
      - Customer Location
  /customers:
    delete:
      description: Delete Customer
      parameters:
      - description: Customer ID
        in: query
        name: customer_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
      summary: Delete Customer
      tags:
      - Customer
    get:
      description: Retrieve Customer
      parameters:
      - description: Customer Name
        in: query
        name: customer_name
        type: string
      - description: Customer ID
        in: query
        name: customer_id
        type: string
      - description: Created By
        in: query
        name: created_by
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order_by
        type: string
      - description: Select Column
        in: query
        name: select_column
        type: string
      - description: Set Limit
        in: query
        name: set_limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
      summary: Retrieve Customer
      tags:
      - Customer
    patch:
      consumes:
      - application/json
      description: Update Customer
      parameters:
      - description: Customer Data
        in: body
        name: Customer
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCustomer'
      - description: Customer ID
        in: query
        name: customer_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
      summary: Update Customer
      tags:
      - Customer
    post:
      consumes:
      - application/json
      description: Add    Customer
      parameters:
      - description: Customer Data
        in: body
        name: Customer
        required: true
        schema:
          $ref: '#/definitions/models.CreateCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
      summary: Create Customer
      tags:
      - Customer
  /email-domains:
    delete:
      description: Delete Email Domain
      parameters:
      - description: Email Domain Id
        in: query
        name: email_domain_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmailDomain'
      summary: Delete Email Domain
      tags:
      - Email Domain
    get:
      description: Retrieve Email Domain
      parameters:
      - description: Domain
        in: query
        name: domain
        type: string
      - description: Email Domain Id
        in: query
        name: email_domain_id
        type: string
      - description: Customer Location
        in: query
        name: customer_location
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmailDomain'
      summary: Retrieve   Email Domain
      tags:
      - Email Domain
    patch:
      consumes:
      - application/json
      description: Update Email Domain
      parameters:
      - description: Email Domain
        in: body
        name: EmailDomain
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/models.UpdateEmailDomain'
          - properties:
              account:
                $ref: '#/definitions/models.UpdateAccount'
              email_account:
                $ref: '#/definitions/models.UpdateEmailAccount'
            type: object
      - description: Email Domain ID
        in: query
        name: email_domain_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Customer'
      summary: Update Email Domain
      tags:
      - Email Domain
    post:
      consumes:
      - application/json
      description: Add    Email Domain
      parameters:
      - description: EmailDomain Data
        in: body
        name: EmailDomain
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/models.CreateEmailDomain'
          - properties:
              account:
                $ref: '#/definitions/models.CreateAccount'
              email_account:
                $ref: '#/definitions/models.CreateEmailAccount'
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmailDomain'
      summary: Create Email Domain
      tags:
      - Email Domain
...
schemes:
- http
swagger: "2.0"